temporarily support space creation using service account (#513)

+++ controller/space.go
@@ -11,6 +11,7 @@ import (
 	"github.com/fabric8-services/fabric8-auth/log"
 	"github.com/fabric8-services/fabric8-auth/login"
 	"github.com/fabric8-services/fabric8-auth/space"
+	"github.com/fabric8-services/fabric8-auth/token"
 
 	"github.com/goadesign/goa"
 	"github.com/satori/go.uuid"
@@ -47,31 +48,55 @@ func NewSpaceController(service *goa.Service, app application.Application, confi
 
 // Create runs the create action.
 func (c *SpaceController) Create(ctx *app.CreateSpaceContext) error {
-	currentIdentity, err := login.LoadContextIdentityIfNotDeprovisioned(ctx, c.app)
-	if err != nil {
-		return jsonapi.JSONErrorResponse(ctx, err)
+
+	isMigration := false
+	if token.IsSpecificServiceAccount(ctx, "space-migration") {
+		isMigration = true
 	}
 
-	// Create keycloak resource for this space
-	resource, err := c.resourceManager.CreateResource(ctx, ctx.RequestData, ctx.SpaceID.String(), spaceResourceType, nil, &scopes, currentIdentity.ID.String())
-	if err != nil {
-		return jsonapi.JSONErrorResponse(ctx, err)
-	}
-	spaceResource := &space.Resource{
-		ResourceID:   resource.ResourceID,
-		PolicyID:     resource.PolicyID,
-		PermissionID: resource.PermissionID,
-		SpaceID:      ctx.SpaceID,
-		OwnerID:      currentIdentity.ID,
+	var currentIdentity *account.Identity
+	var err error
+	var spaceResource *space.Resource
+	var resource *auth.Resource
+
+	if isMigration {
+		currentIdentity = &account.Identity{ID: *ctx.Creator}
+	} else {
+		currentIdentity, err = login.LoadContextIdentityIfNotDeprovisioned(ctx, c.app)
+		if err != nil {
+			return jsonapi.JSONErrorResponse(ctx, err)
+		}
 	}
 
-	err = transaction.Transactional(c.app, func(tr transaction.TransactionalResources) error {
-		// Create space resource which will represent the keycloak resource associated with this space
-		_, err = tr.SpaceResources().Create(ctx, spaceResource)
-		return err
-	})
-	if err != nil {
-		return jsonapi.JSONErrorResponse(ctx, err)
+	if !isMigration {
+
+		// Create keycloak resource for this space
+		resource, err = c.resourceManager.CreateResource(ctx, ctx.RequestData, ctx.SpaceID.String(), spaceResourceType, nil, &scopes, currentIdentity.ID.String())
+		if err != nil {
+			return jsonapi.JSONErrorResponse(ctx, err)
+		}
+		spaceResource = &space.Resource{
+			ResourceID:   resource.ResourceID,
+			PolicyID:     resource.PolicyID,
+			PermissionID: resource.PermissionID,
+			SpaceID:      ctx.SpaceID,
+			OwnerID:      currentIdentity.ID,
+		}
+
+		err = transaction.Transactional(c.app, func(tr transaction.TransactionalResources) error {
+			// Create space resource which will represent the keycloak resource associated with this space
+			_, err = tr.SpaceResources().Create(ctx, spaceResource)
+			return err
+		})
+		if err != nil {
+			return jsonapi.JSONErrorResponse(ctx, err)
+		}
+		log.Debug(ctx, map[string]interface{}{
+			"space_id":      ctx.SpaceID,
+			"resource_id":   resource.ResourceID,
+			"permission_id": resource.PermissionID,
+			"policy_id":     resource.PolicyID,
+		}, "space resource created")
 	}
 
 	// Create AuthZ resource for the space as part of soft migration from deprecated Keycloak AuthZ API to new OSIO AuthZ API
@@ -81,31 +106,35 @@ func (c *SpaceController) Create(ctx *app.CreateSpaceContext) error {
 			"err":      err,
 			"space_id": ctx.SpaceID,
 		}, "unable to register resource for space or assign space admin role to space creator")
-		// Roll back Keycloak resource creation
-		rolBackErr := transaction.Transactional(c.app, func(tr transaction.TransactionalResources) error {
-			return tr.SpaceResources().Delete(ctx, spaceResource.ID)
-		})
-		if rolBackErr != nil {
-			log.Error(ctx, map[string]interface{}{
-				"err":      rolBackErr,
-				"space_id": ctx.SpaceID,
-			}, "unable to roll back space resource creation")
+
+		if !isMigration {
+
+			// Roll back Keycloak resource creation
+			rolBackErr := transaction.Transactional(c.app, func(tr transaction.TransactionalResources) error {
+				return tr.SpaceResources().Delete(ctx, spaceResource.ID)
+			})
+			if rolBackErr != nil {
+				log.Error(ctx, map[string]interface{}{
+					"err":      rolBackErr,
+					"space_id": ctx.SpaceID,
+				}, "unable to roll back space resource creation")
+			}
 		}
 
 		return jsonapi.JSONErrorResponse(ctx, errors.NewInternalError(ctx, err))
 	}
 
-	log.Debug(ctx, map[string]interface{}{
-		"space_id":      ctx.SpaceID,
-		"resource_id":   resource.ResourceID,
-		"permission_id": resource.PermissionID,
-		"policy_id":     resource.PolicyID,
-	}, "space resource created")
-
+	if resource != nil { // for migration = true, resource is nil
+		return ctx.OK(&app.SpaceResource{Data: &app.SpaceResourceData{
+			ResourceID:   resource.ResourceID,
+			PermissionID: resource.PermissionID,
+			PolicyID:     resource.PolicyID,
+		}})
+	}
 	return ctx.OK(&app.SpaceResource{Data: &app.SpaceResourceData{
-		ResourceID:   resource.ResourceID,
-		PermissionID: resource.PermissionID,
-		PolicyID:     resource.PolicyID,
+		ResourceID:   ctx.SpaceID.String(),
+		PermissionID: uuid.Nil.String(),
+		PolicyID:     uuid.Nil.String(),
 	}})
 }
